Update to latest lib-common and keystone-operator
With [1] the keystone-operator introduces another CRD, KeystoneEndpoint,
which from now on handles the registration of endpoints. The
KeystoneService CRD is now only used to register the service and service
user. This allows to create the service in a different controller/code
path from the endpoints. Service operators use this to e.g. create the
KeystoneService in the top level, register the endpoints in the API
controller, register endpoint types via different CRs if multiple API
services need to run to serve e.g. public and internal requests.

Also with [2] the keystone external package was moved into the
keystone-operator api sub module. With this is no longer required for
3rd party operators to import the keystone-operator base module and
therefore have dependencies on any keystone-operator internals.

[1] https://github.com/openstack-k8s-operators/keystone-operator/pull/100
[2] https://github.com/openstack-k8s-operators/keystone-operator/pull/90