Add webhook defaulter
Add a defaulter to the webhook. Move "transfers from umbrella
Ironic" that are Spec-to-Spec copying out of controller
reconciler and instead implement them in the defaulter webhook.

Implements the exported func (spec *IronicSpec) Default() function
which can be used to default an IronicSpec from external callers,
for example a webhook in openstack-operator should be able to use
it to default the template (spec) for Ironic.

Also adds 'omitempty' to Conductor and Inspector NetworkAttachments
to avoid 'Invalid value: "null"' issues when not used.

Jira: OSP-22601