Align NetworkAttachments with other operators
For network isolation outgoing traffic, service pods might be
attached to additional networks. To achive this, the
`k8s.v1.cni.cncf.io/networks` annotation needs to be set on the
pods. This adds a networkAttachments parameter to the
CRD where it can be specified to which additional networks the
service pods should be attached to.

Depends-On: openstack-k8s-operators/lib-common#133