Unhide KeystoneVars
Using the `json:"-" to "hide" the property cause issues,
the generated config map using the values in the map end
up with empty values. This might be caused by the json
serialisation in the webhooks.

Let's just unhide the property, it may come in handy to
expose them if we ever want to run conductor in a separate
OCP cluster.

Jira: osp-23942