Centralize images in the Ironic CRD
This change creates an Images struct in the Ironic spec to centralize
image tags for all ironic services into a single location. Defaults
for this struct come from the operator environment, as before. And
defaults for the sub specs are inherited from the Ironic spec.

There are a few reasons for this change:
- Pxe image tag is repeated across multiple specs
- Managing image tags across a complex structure may become a burden
  for users
- It is not possible to distribute a stock kustomize patch which will
  update conductor images that works with conductor groups

Regarding the last point, I'm going to propose that kustomize patches
be the mechanism for managing service container updates for all
operators, this change makes that possible for ironic-operator. Since
the webhook only handles setting default values on first deploy, a
follow-up change will be required which propagates image tag changes
to the the sub-controllers when the images change.

This change also ensures all image attributes are optional since the
webhook will populate values in all cases. The old IronicDefaults
structure will remain until openstack-operator has been transitioned
to IronicImages.