Refactor type definitions using templates
This refactors type definitions using templates. Benefit of approach
is that we can hide some specs of sub CRs from the top-level Ironic CR,
and decrease size and complexity of the resource definition.

This also replaces defaulting in webhook. The problem with using
webhook is that the copied default is baked in CR and users should
update all fields when they update existing instances.