Refactor ironic config script to common
Move the common init script code for ironic config
to a function in common.sh. Use the function in
individual init scripts for ironic, ironicapi and
ironcconductor.

Also move parts of the pxe-init dnsmasq config to the
conductor init script and re-organize the init containers
to run pxe-init first - this enables reducing the initcontainers
for conductor from 4 to 3.