Update ironicapi and ironicinspector to use service override
Removes creation of routes for ironicapi and ironicinspector as
those get done in the openstack-operator. Via service overrides
the parts of the service can be customized. The service operator
adds annotation to the service which needs to be exposed as a route.

Note: does not cover the routes being created by the ironicconductor
and ironicinspector for the httpboot service when there is no inspection
network

Jira: OSP-26690

Depends-On: https://github.com/openstack-k8s-operators/lib-common/pull/332