Move some validation from kubebuilder annotation to webhook
Right now applying a OpenStackControlPlane CR without any ironic
section fails with:

~~~
$ oc apply -n openstack -f ...
The OpenStackControlPlane "openstack-galera-network-isolation-3replicas" is invalid:
* spec.ironic.template.rpcTransport: Unsupported value: "": supported values: "oslo", "json-rpc"
* spec.ironic.template.ironicConductors: Invalid value: "null": spec.ironic.template.ironicConductors in body must be of type array: "null"
~~~

To allow the OpenstackControlPlane to be deployed without specifying
any ironic there is the need to change the validation/defaulting of
the RPCTransport and IronicConductors.

* RPCTransport
The enum check is moved to the validations webhook for create
and update, default set in the defaulting webhook.

* IronicConductors
IronicConductors not longer a required parameter from kubebuilder
pov, but in the validations webhook checked that length of the
array is > 0