Fix X-Forwarded-Proto in non-TLS deployments
In the httpd container configuration there was an issue that
"X-Forwarded-Proto" header was setting "https" protocol always, no
matter if TLS was enabled in the deployment or not. It caused issues
in neutron server with pagination when TLS was disabled as client was
doing request to the "http://" endoint and got in response
"https://" links to "next" and "prev" pages.

This patch fixes this issue by setting correct protocol for both cases:
TLS and non-TLS.

Related: OSPRH-10632

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>