Fix CrashLoopBackOff, ramdisk-logs no inotifywait
When updating the operators we end up recreating the conductor and
inspector deployments. But the old service images does not yet have
the required command: inotifywait - because of the missing command
the pod enter CrashLoopBackOff state.

To resolve the issue, the OpenstackVersion must be patch to update
the container images.

To workaround this issue, add a check for the inotifywait command,
and in case it does not exist:
 a) Print a Warning
 b) Starta a tail for follow /dev/null to have the container do
    nothing forever. (e.g. until it is restarted with new service
    container images after patching the OpenstackVersion.)

Jira: OSPRH-17520