[OSPRH-19691] Add unit test to pre-commit hook
We clsoed the #597, Add unit test for webhook. I tested the changes
locally. Currently we are looking for unit tests only in api directory.

During CR we decided to increase the coverage of
unit tests so that if we add new unit tests for other functionality that
are in other folders except tests it should run that as well as a
pre-commit hook.

Also we decided to change file Makfile test name from test-api to
unit-test.

This PR address the above issue:
1. Replacing `test-api` with `unit-test` in the Makefile.
2. Adding unit-test coverage for whole project excluding test directory.
3. Adding unit-test in the pre-commit hook.

Signed-off-by: Abhishek Bongale <abhishekbongale@outlook.com>