Ensure rpc_transport is set in ironic.conf
Fixes an issue where `rpc_transport` is not set to `oslo`
even if instance.Spec.RPCTransport is `oslo`. This caused
the service to try to connect using json-rpc, which failed
because json-rpc service service is only exposed when
RPC Transport is json-rpc.

Set rpc_transport to match instance.Spec.RPCTransport.

Jira: OSPRH-19936