CRD changes for graphical console
This change introduces the CRD changes for the feature flag and new
images required for graphical consoles.

This is required to add without any backing implementation so that
openstack-operator can pull in the CRD changes and establish the CSV
entries for overriding the images.

The feature flag will remain false until RHOSO-19 and this will never
work with default antelope images.

Jira: OSPRH-20211